Small memory saving + test data + reorg
Added a delete for no longer used list
Added some more test data (SHOULD NOT BE PUSHED TO MASTER)
Rearranged some of the code to be better organized
